requirements.txt
azure-storage-blob==12.24.1
beautifulsoup4==4.12.3
selenium==4.27.1